Added netcore solution.

+++ src/Reminiscence/Arrays/MemoryArray.cs
@@ -82,7 +82,7 @@ namespace Reminiscence.Arrays
             }
             set
             {
-                long block = (long)System.Math.Floor(idx / _blockSize);
+                long block = (long)System.Math.Floor((double)idx / _blockSize);
                 long localIdx = idx % _blockSize;
                 blocks[block][localIdx] = value;
             }
